added admin route

+++ server/controllers/blogController.js
@@ -1,4 +1,6 @@
 const AllBlogs = require("../models/AllBlogs");
+const isAdmin = require("../utils/checkAdmin");
+const isAuthor = require("../utils/checkAuthor");
 
 const getAllBlogs = async (req, res) => {
   try {
@@ -6,10 +8,17 @@ const getAllBlogs = async (req, res) => {
     if (req.query.category) {
       query.category = req.query.category;
     }
-    const blogs = await AllBlogs.find(query).sort({ published: -1 });
-    return res
-      .status(200)
-      .json({ success: true, total: blogs.length, data: blogs });
+    const blogs = await AllBlogs.find(query)
+      .sort({ published: -1 })
+      .select("-content -_id -__v -authorId");
+    const filteredBlogs = await blogs.filter(
+      (blog) => blog.status === "published"
+    );
+    return res.status(200).json({
+      success: true,
+      total: filteredBlogs.length,
+      data: filteredBlogs,
+    });
   } catch (error) {
     return res.status(500).json({ error: "Server Error" });
   }
@@ -26,9 +35,11 @@ const getBlogById = async (req, res) => {
 
 const getBlogsByAuthorId = async (req, res) => {
   try {
-    const blogs = await AllBlogs.find({ authorId: req.params.authorId }).sort({
-      published: -1,
-    });
+    const blogs = await AllBlogs.find({ authorId: req.params.authorId })
+      .sort({
+        published: -1,
+      })
+      .select("-content -authorId -_id -__v -authorId");
     return res
       .status(200)
       .json({ success: true, total: blogs.length, data: blogs });
@@ -48,12 +59,18 @@ const createBlog = async (req, res) => {
 
 const updateBlog = async (req, res) => {
   try {
+    const Author = await isAuthor(req.user.userId);
+
+    if (!Author) {
+      return res
+        .status(403)
+        .json({ error: "You are not authorized to update this blog" });
+    }
+
     const id = req.params.id;
     const blog = await AllBlogs.findOneAndUpdate({ slug: id }, req.body, {
       new: true,
     });
-
-
 
     if (!blog) {
       return res
@@ -70,8 +87,17 @@ const updateBlog = async (req, res) => {
 
 const deleteBlog = async (req, res) => {
   try {
-    const blog = await AllBlogs.findByIdAndDelete(req.params.id);
-    return res.status(200).json({ success: true, data: blog });
+    const Author = await isAuthor(req.user.userId);
+    const Admin = await isAdmin(req.user.userId);
+
+    if (Author || Admin) {
+      const blog = await AllBlogs.findByIdAndDelete(req.params.id);
+      return res.status(200).json({ success: true, data: blog });
+    }
+
+    return res
+      .status(403)
+      .json({ error: "You are not authorized to delete this blog" });
   } catch (error) {
     return res.status(500).json({ error: "Server Error" });
   }
@@ -79,7 +105,9 @@ const deleteBlog = async (req, res) => {
 
 const getRecentBlogs = async (req, res) => {
   try {
-    const blogs = await AllBlogs.find().sort({ published: -1 }).limit(6);
+    const blogs = await AllBlogs.find({ status: "published" })
+      .sort({ published: -1 })
+      .limit(6);
     return res
       .status(200)
       .json({ success: true, total: blogs.length, data: blogs });
@@ -101,6 +129,39 @@ const getSearchedBlog = async (req, res) => {
   }
 };
 
+const getStats = async (req, res) => {
+  try {
+    const id = req.user.userId;
+    const statsArray = await AllBlogs.aggregate([
+      {
+        $match: { authorId: id },
+      },
+      {
+        $group: {
+          _id: null,
+          total: { $sum: 1 },
+          published: {
+            $sum: {
+              $cond: [{ $eq: ["$status", "published"] }, 1, 0],
+            },
+          },
+          draft: {
+            $sum: {
+              $cond: [{ $eq: ["$status", "draft"] }, 1, 0],
+            },
+          },
+        },
+      },
+    ]);
+
+    const stats = await statsArray[0];
+
+    return res.status(200).json({ success: true, data: stats });
+  } catch (error) {
+    return res.status(500).send({ message: error.message });
+  }
+};
+
 module.exports = {
   getAllBlogs,
   getBlogById,
@@ -110,4 +171,5 @@ module.exports = {
   deleteBlog,
   getRecentBlogs,
   getSearchedBlog,
+  getStats,
 };
